backend-eva/pyproject.toml
[project]
name = "backend-eva"
version = "0.1.0"
description = "Backend para la plataforma de aprendizaje EVA"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=6.0,<7.0",
    "django-ninja>=1.4,<2.0",
    "channels[daphne]>=4.3,<5.0",
    "celery>=5.5,<6.0",
    "redis>=5.0,<6.0",
    "pyjwt>=2.10,<3.0",
    "nh3>=0.2.21",
    "pydantic-settings>=2.7,<3.0",
    "psycopg[binary]>=3.2,<4.0",
    "channels-redis>=4.3,<5.0",
    "django-extensions>=4.1",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-django>=4.8",
    "hypothesis>=6.130",
    "ruff>=0.14",
    "mypy>=1.18",
]
